requirements.txt
Flask==2.2.3
flask-session==0.4.0
gunicorn==20.1.0
pandas==1.5.3
pdfplumber==0.7.6
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.9
google-api-python-client==2.79.0
google-auth-oauthlib==1.0.0
openpyxl==3.1.2
